add sgt_blaze rules for locomotive

+++ src/cards/greattrainrobbery/full_steam.h
@@ -6,7 +6,7 @@
 namespace banggame {
 
     struct effect_full_steam {
-        int value;
+        int8_t value;
         effect_full_steam(int value) : value(value) {}
         
         void on_play(card *origin_card, player *origin);
